Issue5_js読み込みの修正
