+++ app/controllers/items_controller.rb
@@ -13,7 +13,7 @@ class ItemsController < ApplicationController
     @item = Item.new(item_params)
     if @item.valid?
       @item.save
-      redirect_to items_path
+      redirect_to root_path
     else
       render :new
     end
