+++ app/javascript/item_price.js
@@ -1,0 +1,9 @@
+window.addEventListener('load', () => {
+  const priceInput = document.getElementById("item-price");
+   priceInput.addEventListener("input", () => {
+     const addTaxDom = document.getElementById("add-tax-price");
+     addTaxDom.innerHTML = Math.floor(priceInput.value * 0.1 );
+     const addPofitDom = document.getElementById("profit");
+     addPofitDom.innerHTML = Math.floor(priceInput.value - Math.floor(priceInput.value * 0.1 ))
+ })
+});
