+++ app/javascript/packs/application.js
@@ -16,12 +16,4 @@ require("channels")
 // const images = require.context('../images', true)
 // const imagePath = (name) => images(name, true)
 
-window.addEventListener('load', () => {
-  const priceInput = document.getElementById("item-price");
-   priceInput.addEventListener("input", () => {
-     const addTaxDom = document.getElementById("add-tax-price");
-     addTaxDom.innerHTML = Math.floor(priceInput.value * 0.1 );
-     const addPofitDom = document.getElementById("profit");
-     addPofitDom.innerHTML = Math.floor(priceInput.value - Math.floor(priceInput.value * 0.1 ))
- })
-});
+require("item_price");
